testing for data processing
